Write the Code to Play One Round
This deployment added the structure of the game. For the computer choices, a Math.floor(Math.random() * 3) is used to generate three random numbers (0,1,2), then used as an input for a switch statement to return one of the three possible choices. The user input is captured using prompt() then converted to lower case to be case insensitive. Then the comparison is carried out using if statement.