Add for loop to play the game for 5 rounds
The game has now been modified to have 5 rounds with a for loop. The if statement that compares both choices at each round in now included in the loop, once the loop is done, the function (game) goes into another if statement to declare the winner along with the final result.